fix Packages.ps1
Persimmon.Console should not have the lib directory in the nupkg file.
But if we create the nupkg file using the fsproj file, create nupkg file
always contains the lib directory.
So I introduce the template file that generates nuspec file.
Now, we create the nupkg file using the nuspec file.